db and script changes
