+++ runinator-worker/Cargo.toml
@@ -11,6 +11,7 @@ chrono = { version = "0.4", features = ["serde"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 uuid = { version = "1.3", features = ["v4", "serde"] }
+log-panics = "2.1"
 
 runinator-comm = { path = "../runinator-comm" }
 runinator-plugin = { path = "../runinator-plugin" }
@@ -18,4 +19,3 @@ runinator-provider-aws = { path = "../runinator-provider-aws" }
 runinator-provider-sql = { path = "../runinator-provider-sql" }
 runinator-models = { path = "../runinator-models" }
 runinator-utilities = { path = "../runinator-utilities" }
-log-panics = "2.1"
