+++ runinator-ws/Cargo.toml
@@ -12,7 +12,9 @@ clap = { version = "4.5", features = ["derive"] }
 env_logger = "0.11"
 chrono = { version = "0.4", features = ["serde"] }
 uuid = { version = "1.3", features = ["v4"] }
+log-panics = "2.1"
 
 runinator-database = { path = "../runinator-database" }
 runinator-models = { path = "../runinator-models" }
 runinator-comm = { path = "../runinator-comm" }
+runinator-utilities = { path = "../runinator-utilities" }
